Tentative fix for standard gear reset on rare occasions
This may or may not work. The problem is not yet fully understood, but it seems to happen if a player dies and respawns during loss of connection. This causes HandleDisconnect to see the player as alive and fully respawned, therefore saving the default gear equipped by the unit.

I don't know if the player unit is local or remote to the server when this kind of respawn happens, but if it's server-side, this should ensure that the unit will receive the "playerSpawning" flag, as required to skip the gear save from HandleDisconnect.